Fix for bone weights accepting only integer values. Fix for crash on opening character property bone weights for edit. Implementation of character property removal. Fixed display of bone weights in character properties. Fixed character properties table update after loading different project. Added hack with copying of bone weights on setting character property binding to bone weights array in blender generator child.

+++ src/hkxclasses/behavior/hkbboneweightarray.cpp
@@ -19,7 +19,7 @@ hkbBoneWeightArray::hkbBoneWeightArray(HkxFile *parent, long ref, int size)
     refCount++;
 }
 
-void hkbBoneWeightArray::setBoneWeightAt(int index, int value){
+void hkbBoneWeightArray::setBoneWeightAt(int index, qreal value){
     std::lock_guard <std::mutex> guard(mutex);
     (index >= 0 && index < boneWeights.size() && value < static_cast<BehaviorFile *>(getParentFile())->getNumberOfBones()) ? boneWeights[index] = value, getParentFile()->setIsChanged(true) : LogFile::writeToLog(getParentFilename()+": "+getClassname()+": failed to set boneWeights!");
 }
